Reorganize controllers and routes
